requirements.txt
streamlit
google-generativeai
python-dotenv   # To load all our environmental variables
